Control how far to advance the `format` pointer
The pointer that moves through the characters in `format`
can now be precisely controlled (in terms of how many steps
forward it advances).
This is useful because the pointer can now move either 1 step forward
or 2 steps forward depending on the situation.
The number of steps to advance is now dictated by the return value of
`print_special`